requirements.txt
# YOLO26 SeaTrash Requirements

# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0

# Computer Vision
opencv-python>=4.8.0
Pillow>=10.0.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
tqdm>=4.65.0
PyYAML>=6.0
python-dotenv>=1.0.0

# Optional: Advanced Features
# albumentations>=1.3.0  # Advanced augmentation
# tensorboard>=2.13.0    # Training visualization
# wandb>=0.15.0          # Experiment tracking
